Update motor control and WiFi functionality with documentation

+++ MGv2/include/motor_protocol.h
@@ -26,8 +26,8 @@ struct MotorStatus {
     uint8_t errorState;      // Error flags
     int16_t torqueCurrent;   // iq: -2048~2048 → -33A~33A
     int16_t speed;           // dps (degrees per second)
-    int16_t acceleration;    // dps/s (degrees per second squared)
+    int32_t acceleration;    // dps/s (degrees per second squared) - INT32!
     uint16_t encoder;        // 0~16383 (14-bit)
-    int64_t motorAngle;      // 0.01°/LSB (multi-turn cumulative)
+    int64_t motorAngle;      // 0.01°/LSB (single-turn angle, 0-360°×gear_ratio)
     uint32_t timestamp;      // millis() when read
 };
